kchrome/requirements.txt
# requirements for kodachrome runtime

discord.py
python-dotenv

